[ENHANCEMENT] Refactoring sprite generation (#3876)
* accessory files

* scar files

* adding area detection for accessories

* removing unneeded list assignments

* scar pelt lists

* refactoring load into a func

* skin files

* tortie patch files

* pelt files

* eye files

* white marking files

* small fixes

* bugfixing

* more small fixes

* documentation

* moving these to sprites folder

i dinnae know the sprites folder had a dict folder already

* formatter

* doc fix?

* fix bugs

* delete duplicate doc images

* remove unnecessary bit

* Update utility.py

* formatter

* fixing some small inconsistencies

* further optimizing

* setting up multi_sheet detection

this way we don't need to manually change the tuples for these, instead it just compiles a list based on the data itself

* updating docs

* documentation

* updating pose references and lineart handling

* fixing april fools

* remove unused dict entries

* remove unused dict entries

* documentation regarding adding new poses

* typo fix

* add protection against removed poses

now, if a save file includes an invalid pose name, the default pose will be used instead. this makes it easy for modders to remove poses without having to mess with save conversion.

* documentation regarding pose removal

* doc additions

* Update adding-new-sprites.md