Newborn kitten sprite variants (#4020)
* potatoes in progress

they're a teensy bit temperamental rn..

* potatoes work?

still a wip - white and tortie patches, as well as smoke, sokoke and speckled colours, need to be finished.

* changed starfish

 to match requested adjustments

* lowkey forgot the lineart variants

* potatoes on top!

also adjusted the code and singlecolour spritesheet to account for this.

* getting through the spritesheets :)

thank you artist devs for the help!

* almost all done!

thank you again artist devs for helping :D

* white patches!

* save conversion

* potatoes all done, i think

many thanks to flamedash, scribblecrumb, and eureka from the dev team for helping complete this pr <3

* updating to new json system

* fix newborn sprite loading

---------

Co-authored-by: scribblecrumb <redacted>
Co-authored-by: scribblecrumb <redacted>