[FEATURE] New Patrol Tags: all_mentored and app{index}_mentored (#3984)
* Create A 'has_mentor' Check

Creates two different tags: "app_mentored" (which checks if all apprentices thats within the patrol_apprentices has a mentor) and "app{index}_mentored" (which checks if the apprentice at a specific number within the app list has a mentor).

* Update (Patrol) Tag Filtering

Updates event_filtering to check if the apprentices have a mentor

* Update Docs & Schema

Updates patrol document to include the desc. of both tags, and update the patrol schema with all the new patrol tags.

* Update (Most) Beach Patrols

If any patrol mentioned mentor(s) which indicated ownership from cat(s) in the patrol, it was tagged.

* Tag (Most) Desert Patrols

If any patrol mentioned mentor(s) which indicated ownership from cat(s) in the patrol, it was tagged.

* Tag (Most) Forest Patrols

If any patrol mentioned mentor(s) which indicated ownership from cat(s) in the patrol, it was tagged.

* Tag (Most) General Patrols

If any patrol mentioned mentor(s) which indicated ownership from cat(s) in the patrol, it was tagged.

* Tag (Most) Mountainous Patrols

If any patrol mentioned mentor(s) which indicated ownership from cat(s) in the patrol, it was tagged.

* Tag (Most) Plains Patrols

If any patrol mentioned mentor(s) which indicated ownership from cat(s) in the patrol, it was tagged.

* Tag (Most) Wetlands Patrols

If any patrol mentioned mentor(s) which indicated ownership from cat(s) in the patrol, it was tagged.

* Fix Tests Failing

Fixes some formatting errors that were causing a unit test to fail

* Fix/Apply Requested Changes

Apply the changes requested by reviewers - makes the code more streamlined and simplified.

* Fix Test Fail

Once again, fix my formatting in order for the test to run.

* Update With Requested Changes

Update the has_mentor tag in the function event_for_tags to mentor_tags_fulfilled.

* Fix a Missed has_mentor Call

missed a has_mentor whilst rushing to fix it :3