[FEATURE] Afterlife temperaments (#4043)
* adding afterlife classes and calc-ing temper

* various fixes

* allow instructors

* fix afterlives in tests

* remove unneeded statement

silly scribble, this guy already gets added during cat load

* test fix

* docstrings and attribute renames

* use floor division instead of division

avoids float rounding issues

* readability pass

* switch to a set of cat IDs

now we gather a set of IDs for each cat who influences the afterlife and use it's len for the average calcs. this makes it easier to tell if a cat was already included or not.

* fix accidental rename

* fix guide influence

* explanatory comment

* avoiding roundoff erros

* doc fix

* FIX: feedback changes

* address feedback

* formatter

* rename func